Add 'tsort' utility.
NOTE: There is a discrepancy between the output in the POSIX docs, for the
given POSIX sample, and our output.  It is unclear how to treat items that are
not directly linked to other items... there appears to be some secondary
input-order-based ordering going on that warrant investigation.

We appear to do the right thing when all items are linked.

Signed-off-by: Jeff Garzik <jgarzik@redhat.com>